faas-santa-detector/function/requirements.txt
tensorflow
Pillow
matplotlib
keras
